requirements.txt
django==3.0.7
djangorestframework==3.11.0
psycopg2>=2.7.5,<2.8.0
Pillow==7.1.2

flake8>=3.6.0,<3.7.0